Fixed bug in case file length exceeds max int
git-svn-id: https://svn.code.sf.net/p/openatfx/code/trunk@345 3fec9929-6ec6-4949-bc57-c3a92d7ac694

+++ openatfx/src/main/java/de/rechner/openatfx/util/BufferedRandomAccessFile.java
@@ -63,9 +63,9 @@ public class BufferedRandomAccessFile extends RandomAccessFile {
                 continue;
             }
             int toread = stop - idx;
-            int available = (int) (fileLength - filePos);
+            long available = (fileLength - filePos);
             if (toread > available)
-                toread = available;
+                toread = (int) available;
             int availableInBuffer = bufferSize - readAtIdx;
             if (toread > availableInBuffer)
                 toread = availableInBuffer;
@@ -74,7 +74,7 @@ public class BufferedRandomAccessFile extends RandomAccessFile {
             filePos += toread;
             // check if end of file is reached
             if (filePos == fileLength)
-                return toread-off;
+                return toread - off;
         }
         return idx - off;
     }
